Seperate xpath parser to a single file.
